Fix for Instrument Agent Instance association types
Was seeing an issue with editing an InstrumentAgentInstance's
associations where the resource_identifier field looked incorrect and
therefore I couldn't substitute the correct variable in the unfilled
request params. Judging by the other Prepare* objects, such as
PlatformDevice, I think the types should reflect the OTHER side of the
association.