TEST - Fixes broken smoke tests.
The following changes cause the granule to instantiate an empty stream
definition. The prior behavior was that the field was initialized to None and
the parsers expect Nones for granules without a stream definition (retrieve) but
there were stream definition objects, just without the fields filled in.

By changing the type from !StreamDefinition to {}, the initialized field
should be an empty dict, which will evaluate as False in a boolean expression.

The changes:
https://github.com/ooici/pyon/compare/9b4a656b978ad39d8e5f67fe66b2913e99569a30...ad73ad95628b0c5cc646420dbd50fd7df934bb0d

The code that fails:
https://github.com/lukecampbell/coi-services/blob/7cc027c8ecdb17ba2c1ef6eb6d8b02fb1198a09a/ion/services/dm/utility/granule/record_dictionary.py#L179-180