Adds service method to data prodc mgmt svc: add_parameter_to_data_product
I needed a service method at the SA level to maintain parameter consistency
across the associated resources with the data product: the parameter
dictionary, the stream definition and the dataset / coverage model. Originally
this was just going to be in dataset management but the code required going up
the chain to the SA level and then back down. This approach is a bit more
intuitive and straight forward.