Adds a dataset.xml template
So now we have two templates. We have a full catalog template that includes a
partial and a container to produce a full datasets.xml. We also have a
dataset.xml that is only the partial, can be used to prodce a single dataset
tag in the catalog. The latter will be used when a data product is changed.